Clean up API tests
